[hmac,dv] Support new HMAC IP in DV
This incorporates initial changes in the HMAC DV scoreboard
and DV test sequences to support the new HMAC IP with configurable
extended digest size and configurable key length. This also changes
when the predicted digest is computed, and retains previous prediction so long
as the hmac/sha-2 operation is not started. Smoke test passes for all
digest modes and rest of regression tests pass for 256-bit functionalities.

Signed-off-by: Ghada Dessouky <gdessouky@google.com>