[usbdev,sival] Implement usbdev_aon_wake_reset
This test is a bit tricky because it requires to manually suspend
the device and issue a reset while in suspended state. On linux,
suspended a device is not trivial and requires poking at various
sysfs files, and resetting a suspended device is not possible.
Instead, the approach used here is to directly talk to the parent
hub of the device and issue suspend/reset requests directly. This
is a bit risky in that it assumes that linux won't mess with the
hub state while we do that but that seems to work. As stated
in the tests, we could make this actually safe by manually unbinding
the device from the driver first to ensure no intereference but
again this requires some sysfs magic and more persmissions.

A side effect of this is that the test requires access not only to
the USB device under test but also its parent hub. This access is
not available currently in CI so the test is tagged as manual for now.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit b3b243403254305dc65ec7ecb8f0fd65d1a14732)