[opentitantool] Fix crippled I2C TPM support
Is seems that PR #21289 introduced a bug that prevents `opentitantool
i2c tpm ...` from working at all.

This CL restores the proper use of `Any<>` to allow passing the I2C
driver object to the TPM command handler in the way that it expects.

Also, a logic condition was accidentally reversed, making the new
--gsc-ready flag ineffective for I2C.

Change-Id: I2b791de25666876c49dccccd83bdbc9369e608be
Signed-off-by: Jes B. Klinke <jbk@chromium.org>