[jtag_dmi_monitor,dv] Make monitor pickier about the DMI sequence
It turns out that seeing a DmiOpOk JTAG response isn't *quite* enough
to conclude that the DMI is not busy. If you're a bit unlucky, it can
only realise that it's busy at the same time as it responds with
DmiOpOk the first time.

This change makes the monitor "safer" in this setting. It will only
conclude that a requested DMI operation has been sent after it has
seen another response since the original request.

Of course, there's still a question about what to do if we see a
request immediately after a request. The code here goes for the easy
option and fails instantly. We can fix the drivers to ensure they
don't do that!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>