[ipgen,clkmgr] Handle earlgrey and englishbreakfast differences
The earlgrey and englishbreakfast CSR differences mean they have a
different number of clock measurement units. The ideal way to handle
this difference is just generating the dif code, and we intend to do
that with multi-top. However, for the time being we can rely on
macro trickery, just generating C code for a specific measurement if
there is a #define for the corresponding enable CSR offset.

The rust side has a similar problem, but this simply removes the
clkmgr.rs and extclk.rs code from opentitanlib, since it is only used
for enabling external clock in a specific test.

Part of #8440

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>