[ipgen,clkmgr] Handle earlgrey and englishbreakfast differences
The earlgrey and englishbreakfast CSR differences mean they have a
different number of clock measurement units. The ideal way to handle
this difference is just generating the dif code, and we intend to do
that with multi-top. However, for the time being we can rely on
macro trickery, just generating code for a specific measurement if
there is a #define for the corresponding enable CSR offset.

Part of #8440

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>