[spi_device] Fix up TPM synchronization between domains
Fix up synchronization between various domains for the TPM mode.

Synchronize the assertion of the TPM CSB "reset" with the sys_clk
domain, not just its release. CSB is active while sys_clk logic is
active, so prim_rst_sync isn't enough to safely synchronize the reset on
its own.

Replace the RdFIFO's FIFO interface on the SPI clock side with a simpler
memory interface. Remove the pointer tracking across clock domains, and
replace it with a "go" signal from the sys_clk domain, representing that
enough data is available in the Read "FIFO" buffer.

Add an interrupt that triggers when commands targeting the Read FIFO
end, regardless of the outcome of the transaction. Report whether the
transaction failed in TPM_STATUS. Adjust the "rdfifo data dropped"
interrupt to only trigger when a TL-UL write is not accepted into the
SRAM due to a read command not being active.

Document the clock ratio requirement for core clock vs. SPI clock.

Signed-off-by: Alexander Williams <awill@opentitan.org>