[jtag_dmi_reg_frontdoor] Send an extra JTAG read after a DMI request
This fixes a rather confusing behaviour of our DMI over JTAG
implementation. If the JTAG clock happens to be running much faster
than the main clock, we might respond with a spurious DmiOpOk to the
first transaction after a request, even if the request is still being
handled.

Gah!

Make the driver careful to send a spare JTAG read to ensure we really
are waiting until the operation is finished.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>