[ci] Avoid running duplicate hyperdebug-based tests
At the moment, the sival job runs only `cw310_sival` jobs but not
`cw310_sival_rom_ext` which is not right. On the other hand,
the hyperdebug job runs all `hyper310` tests and that includes a
number of `cw310_sival` and `cw310_sival_rom_ext` tests so we run
some tests twice.

This commit changes removes the existing hyper310 job and split the
sival job into two subsets:
- sival_rom_ext tests,
- sival tests that do not have a corresponding sival_rom_ext tests.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>