[dv] Change implementation of special mubi access modes
It turns out that UVM_PREDICT_DIRECT cannot be used in certain cases
where the register is accessed via a front-door access (e.g. via the
write/read member functions), because in those cases the register is
flagged as being busy. This causes the UVM_PREDICT_DIRECT function to
return a non-ok status, and it reports a warning every time this
happens, leading to test failure.

That means that we have to use UVM_PREDICT_WRITE/READ instead. Since
that means that the access policies are enforced at the lowest
implementation level in the uvm_reg_field class, we can't use access
policies such as W1C for mubi fields.

This patch therefore adds an additional field called "mubi_access" to the
dv_base_reg_field class. In case the field represents a mubi
with a special access policy, this field encodes the intended access
policy, while the uvm access filed is just set to "RW".

This way, we can resolve the correct behavior in the do_predict()
function of the derived dv_base_reg_field class.

Signed-off-by: Michael Schaffner <msf@opentitan.org>