[dv] Case 'CompareCtrConsistency' split cases '*Done' & '*Waiting'
The sequence in case 'CompareCtrConsistency' tried to inject errors in
`rom_ctrl_compare.addr_q`. However, it didn't handle injecting errors
when the FSM of `rom_ctrl_compare` was in `state_q=Waiting` correctly.

The case has been split between `CompareCtrConsistencyWaiting` and
`CompareCtrConsistencyDone` and each of them handle error injection in
states `Waiting` and `Done` respectively.

For `CompareCtrConsistencyWaiting` the delay before injecting an error
has also been reduced by a fifth to increase the chances of the FSM in
`Waiting` state.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>