[fpga_cw310_sival_rom_ext] Use fake-signed tests
Change the generic ROM_EXT target to be signed by the fake prod key, so
it may be used in the prod life cycle state.

Change the fpga_cw310_sival_rom_ext execution environment to use the
generic ROM_EXT with a sival OTP configuration.

This fixes a key mismatch due to earlier changes.

Signed-off-by: Alexander Williams <awill@opentitan.org>