[fpga_cw310_sival_rom_ext] Fix key mismatch
Change the generic ROM_EXT target to be signed by the fake prod key, so
it may be used in the prod life cycle state.

Change the fpga_cw310_sival_rom_ext execution environment to optionally
use the generic ROM_EXT with a sival OTP configuration. A build setting
allows it to be built with the real keys, however.

This fixes a key mismatch due to earlier changes.

Signed-off-by: Alexander Williams <awill@opentitan.org>