[quality] Change the way rustfmt_check and rustfmt_fix work
Previously, rustfmt_fix was calling a custom script but actually,
while the intent was to call the official `rustfmt`, it was calling
`@rules_rust//tools/rustfmt` which is a different thing: it's a driver
that does bazel queries to find out which files to change. The idea
being this driver is that one can already just run
```
bazel run @rules_rust//tools/rustfmt
```
and it will fix all the local rust files in the repository (and
ignores the ones tagged as `norustfmt`). This commit changes
`rustfmt_fix` to essentially do that.

Unfortunately, `@rules_rust//tools/rustfmt` does do check-only
in the upstream version. This commit relies on a custom modification
of rules_rust so that the driver accepts a `--check` option. This way,
rustfmt_check and rustfmt_fix consider exactly the same set of files.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>