[entropy_src/rtl] Keep applying fw_ov_rd_fifo_overflow instead of pulsing
This commit applies the fw_ov_rd_fifo_overflow signal in case of an
overflow. It used to only send a pulse through the register but now
the signal is applied until the FIFO has been emptied once and is
receiving contiguous data again.
This commit also aligns the documentation for the FW_OV_RD_FIFO_OVERFLOW
register. The register is now not clearable by software. It instead
is cleared by hardware.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>