[sival, rv_dm] implement rv_dm_mem_access sival test
This mostly reflects the corresponding DV test, except that instead of
backdooring ROM writes, it just checks if the ROM region reads the expected
value from the ROM binary.

The number of accesses per region is reduced from 128 in DV to 32 to make
the test complete in reasoanble time, given JTAG access is slow.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>