[rtl] Weaken types in mubi*_test_invalid
The existing types were a bit strong. Xcelium didn't seem to mind, but
VCS (correctly) fails to run:

  mubi4_t val;
  `DV_CHECK_STD_RANDOMIZE_WITH_FATAL(val, mubi4_test_invalid(val);)

because it refuses to pick a value for val other than one of the
enumeration values. But they are both valid!

One option would be to use casts to force a bad value into a mubi4_t,
like this:

    mubi4_t val;
    logic [3:0] raw_val;
    `DV_CHECK_STD_RANDOMIZE_WITH_FATAL(
      raw_val,
      raw_val != MuBi4True && raw_val != MuBi4False;)
    val = mubi4_t'(raw_val);

But that seems a bit convoluted. And it requires us to inline the
contents of mubi4_test_invalid instead of using the function. How
silly!

Weaken the type so that the function can be used in the obvious way:

    logic [3:0] val;
    `DV_CHECK_STD_RANDOMIZE_WITH_FATAL(val, mubi4_test_invalid(val);)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>