[bazel] unifty cw310_params, cw310/hyper310_jtag_params
The functionality is integrated to cw310_params as extra parameters.
The intention is that one can choose to enable JTAG without having
to clear the bitstream.

For backward compatibility of the rules, *_jtag_params are set to clear
the bitstream.

If test_harness is specified, then `--bitstream={bitstream}` is automatically
added to test_cmd.

Signed-off-by: Gary Guo <gary.guo@lowrisc.org>