[rom_ctrl,dv] Remove incorrect cast in get_invalid_mubi4
The previous code doesn't actually work in VCS. The problem is that
the VCS constraint solver only looks for known values in enumeration
types. But mubi4_test_invalid(mubi4_t'(...)) will then check:

  mubi4_test_invalid(MuBi4True)   (false)
  mubi4_test_invalid(MuBi4False)  (false)

and conclude that there is no way to do the randomisation. Oops!
Weaken the types to match what we actually want ("Give me a 4-bit
number that isn't a known MuBi4")

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>