[sw/i2c_testutils] Take I2C instance ID into account for pins
The previous `kI2cPlatformPins` array mapped to pins solely based on the
platform.  This works for CW310 (with PMOD and HyperDebug) where I2C is
mapped to one pair of pins for all I2C instances.  In DV, however, each
I2C instance has its own agent, which is connected to its own pair of
(muxed) pins.  Coincidentally, when the I2C instance ID (i.e., 0 to 2)
is (incorrectly) used to index the `kI2cPlatformPins` array, that would
result in the correct pair of pins for DV.  As we're adding a fourth I2C
instance, however, this will no longer work.

This commit takes the I2C instance ID into account when mapping to pins.
The result is the same as before on the condition that the platform
argument to a call of `i2c_testutils_select_pinmux` in
`i2c_device_tx_rx_test.c` gets fixed (which this commit also does), but
now it will be possible to add a fourth I2C instance while keeping the
three platforms.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>