[bazel] Rework bitstream clearing in tests
At the moment, some tests manually clear the bitstream *before* the
tests but really we ought to be doing the opposite which is to
clear the bitstream *after* the test if it touches the OTP. This way,
such a test can be safely followed by a "normal" tests.

This commit introduces the notion of a post_test_{harness,cmd} that
is run after the normal test_harness (on success and failure).
The fpga310 parameters are modified so that they take an extra
"clear_bitstream" argument that uses this post test command to
clear the bitstream if requested.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 10901ed3c654b7e5dac96db12817c33e9686ee60)