[kmac] Rework fifo_empty interrupt, change type to status
The type of this interrupt is changed from event to status. To make the
interrupt more usable for software, it only fires if the message FIFO
has run full previously (e.g., if the SHA3 core is currently busy
or if the KMAC block is waiting for entropy from EDN). Under normal
circumstances, the hardware empties the FIFO much faster than software
filling it, so there is no point in interrupting software to tell it
to run faster.

Also, the interrupt can now only fire if software can actually write
the message FIFO.

This resolves lowRISC/OpenTitan#21049.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>