[hw,spi_host,i2c,soc_proxy] Fix CDC issue fot lsio_triggers
This PR flops the LSIO triggers at the source IP with their respective
peripheral clock and then uses a synchronizer cell at the soc_proxy
IP with the fast SoC clock to fix a CDC issue between both domains.

The LSIO trigger of the I2C IP is already flopped.

Signed-off-by: Robert Schilling <rschilling@rivosinc.com>