[i2c] Increase maximum `FifoDepth` from 127 to 4095
This is a prerequisite for increasing the depth of some of the FIFOs
beyond 127.

This implies that the four FIFO levels can no longer be exposed to SW in
the single 32-bit `FIFO_STATUS` register.  Instead, each of the two
host-mode FIFO levels is now a 12-bit field in the `HOST_FIFO_STATUS`
register, and each of the two target-mode FIFO levels is now a 12-bit
field in the `TARGET_FIFO_STATUS` register.

Co-authored-by: Adrian Lees <a.lees@lowrisc.org>
Signed-off-by: Andreas Kurth <adk@lowrisc.org>