[i2c] Add AcqFifoDepth parameter
This commit splits the depth of the ACQ FIFO from the `FifoDepth`
parameter into a separate, new parameter named `AcqFifoDepth`.  This
allows instantiators to make the ACQ FIFO of any `i2c` instance deeper
or shallower than the other FIFOs.

This commit also makes initial DV changes to cover the new parameter,
although for now `AcqFifoDepth` in DV is simply set to a fixed value
that is different from that of `FifoDepth`.  Later work should cover
different values of the parameter.

The three `i2c` instances in Earlgrey have the new parameter set to 64,
which is equal to the value of the `FifoDepth` parameter, for now.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>