[i2c] Add AcqFifoDepth parameter
This commit splits the depth of the ACQ FIFO from the `FifoDepth`
parameter into a separate, new parameter named `AcqFifoDepth`.  This
allows instantiators making the ACQ FIFO of all `i2c` instances deeper
or shallower than the other FIFOs.

For now, the new parameter is set to 64, which is equal to the value of
the `FifoDepth` parameter.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>