[dv] Read buffer watermark interrupt fix
The spi-device scoreboard was updating the readbuffer watermark (and also read
flip) within a DV_SPINWAIT in the task process_read_command.
The DV_SPINWAIT macro was being killed after CSB was going high since
the spi monitor was setting item.mon_item_complete.

That was killing the thread in charge of predicting watermark and flip
interrupts.

The threads updating watermark and flip interrupts have now been moved
to a separate fork-join_none and are triggered upon a
systemverilog event.

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>