[pentest] Add support for AES SCA measurements on chip
On the FPGA, the AES block automatically raises the trigger allowing
to perform SCA analysis. As this is not available on the chip, this
PR extends the AES SCA code to manually handling the trigger.

The arming of the trigger is implemented in the
`sca_call_and_sleep_trigger` to keep the trigger window as narrow
as possible (this reduces the number of samples needed and helps
to easier setup the SCA gear).

The host code is located in lowRISC/ot-sca#329

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>