[entropy_src/dv] Fix CS AES Halt agent configuration
A visual inspection of the CSRNG block reveales that the maximum latency
for getting a ACK on the CS AES Halt interface is actually 48 clock
cycles (the Update command takes 3 AES-256 blocks).

In addition, the agent was missing the `zero_delays = 0` configuration
parameter. As a result, any min/max delay parameters were ignored and
the agent always ACKed after a single clock cycle.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>