[opentitantool] Remove eprint during legacy_rescue
When OTT is used in scripts, the stderr is typically mapped to stdout.
If we print something to stderr for every progress iteration, is breaks
the progress bar updating in place, and instead each progress bar update
will print on a newline instead of replacing itself.

With this change, the progress bar can update inplace.

Change-Id: I36de1bb2084432a41e6d072a5f1e851639143dba
Signed-off-by: Jett Rink <jettrink@google.com>