[kmac, sha3] Add REQ/ACK interface to delay Keccak operations
This interface is required to be able to properly delay operations of
the SHA3-based conditioner inside the ENTROPY_SRC block until CSRNG
confirms its internal AES core is halted.

Previously, it has been tried to handle this as part of the main FSM
of ENTROPY_SRC outside of the SHA3 core. However, this is not sufficient
as in the case of ENTROPY_SRC, most SHA3 operations are actually
triggered by the sha3pad instance inside the SHA3 core.

This is related to lowRISC/OpenTitan#17941.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>