[sw,i2c_testutils] Fix maximum chunk size in `i2c_testutils_read`
Prior to this commit, `chunk` could exceed one byte if
`I2C_PARAM_FIFO_DEPTH` is larger than 255 (which it isn't in the
current HW parametrization).  Such a chunk value couldn't be transferred
in a single byte.  This commit fixes the code so that it works with the
currently allowed maximum value of `I2C_PARAM_FIFO_DEPTH` (2^12-1).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>