[i2c/rtl] Remove `FifoDepth` parameter from `i2c_core`
The `FifoDepth` parameter is defined in `i2c.hjson` and its value is
reflected in the autogenerated `i2c_reg_pkg`.  Prior to this commit,
`i2c_core` defined `FifoDepth` in its parameter list and the value
defined *there* got used in RTL.  This problem did not have consequences
because the value assigned in `i2c_core` was the same as in `i2c.hjson`.
This commit fixes the problem by importing `FifoDepth` from
`i2c_reg_pkg`.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>