[usbdev] Introduce software resets for buffer FIFOs
As a contingency against failures or the need to reconfigure
the USB device completely software resets for the FIFO are introduced.
Using only the block-level reset could be slow or in some
implementations involve resetting of other IP blocks.
Software resets must be used only when the USB device is disconnected
from the USB because the buffers specified within the Available
SETUP and Available OUT FIFOs may be used at any time.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>