[i2c] Increase depth of ACQ FIFO to 268 bytes
This allows I2C to absorb a max-length SMBus Block Write, which consists
of 1 command code byte, 1 byte defining the number of data bytes in the
block write, up to 255 data bytes, and 1 optional Packet Error Code
(PEC) byte; thus up to 258 bytes in total (from SMBus Specification
version 3.2, section 6.5.7).  Our I2C HW IP module additionally puts the
byte after the Start condition (which contains the address and the R/Wn
bit) and a byte for the Stop condition into the FIFO; thus the total
increases to 260 bytes.

To account for any oversights we may have made in the analysis of the
protocol, we add an extra 8 bytes as margin, thus 268 bytes.

Note that this is larger than required by MCTP, which would be 75 bytes:
- "The size of a transmission unit is defined as the size of the packet
  payload that is carried in an MCTP packet.  The [...] minimum
  transmission unit size for MCTP is 64 bytes.  A message terminus that
  supports MCTP control messages shall always accept valid packets that
  have a transmission unit equal to or less than the baseline
  transmission unit." (from MCTP Base Specification version 1.3.0,
  section 8.3)
- MCTP packets have a 4 byte header followed by the packet payload and
  they are encapsulated between a header and a tailer that is specific
  to the physical medium (from MCTP Base Specification version 1.3.0,
  section 8.1).  For MCTP on SMBus, header and tailer consist of 4 and 1
  byte, respectively (see above).
- Thus 4 + 4 + 64 + 1 = 73 bytes plus two bytes for Start and Stop
  (based on our HW IP module; see above) gives 75 bytes.

If the ACQ FIFO cannot absorb a full SMBus Block Write or MCTP control
message, real-time constraints to process bytes from the ACQ FIFO are
imposed on SW controlling the I2C HW module.  The maximum allowed delays
depend on the protocol:
- SMBus allows target devices to stretch the clock by up to 25 ms per
  message (t_LOW:CEXT in Table 14, 15, and 16 of the SMBus Specification
  version 3.2).
- MCTP allows target devices to stretch the clock by at most 250 us per
  byte (from MCTP SMBbus/I2C Transport Binding Specification version
  1.2.0, section 6.18).

In order to not impose such real-time constraints on SW for top-level
designs that use I2C with SMBus or MCTP, this commit makes the ACQ FIFO
sufficiently deep that any max-length transaction can be absorbed.  If
SW doesn't drain the ACQ FIFO once it's full before the next message,
that message should get NACK'd and the controller has to retry later --
but protocol compliance holds and the I2C lines are not blocked.
Currently our I2C HW IP module doesn't implement this NACKing, but this
will be resolved in a separate commit (see the PR of this commit for
details and links).

Signed-off-by: Andreas Kurth <adk@lowrisc.org>