[entropy_src] Fix FIFO controls, move drop point to before postht FIFO
This commit cleans up and documents the control signals for the main
FIFOs of the pipeline including the esrng, esbit, postht, precon and
esfinal FIFOs. Most of the changes simplify the code but don't alter
the behavior of the design as the used FIFO primitives already
implement the logic to not accept pushes when full internally.

However, there are some important changes that are necessary:
1. The esrng FIFO handles no backpressure anymore. This wasn't spec
   compliant.
2. The sample drop point in case of backpressure is moved to after the
   health tests in case we are not in firmware overide insert mode
   and the window counter controls are updated. This means
   in case of backpressure, samples are tested but they're not pushed
   into the postht FIFO (or the esbit FIFO in case of single-bit mode).
   The window timer doesn't increment to keep the number of samples
   ending up in the conditioner fixed, independent of backpressure.
   This is required by the spec.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>