[entropy_src] Add FIFO to deal with backpressure from the conditioner
This commit adds a 32-bit wide distribution FIFO of configurable depth.
The FIFO is added between the postht FIFO, the observe FIFO, the bypass
FIFO and the precon FIFO. Its main purpose is to buffer entropy bits
while the conditioner is busy such that we don't have to drop entropy
bits from the hardware pipeline.

Dropping entropy bits is not a big issue per se during normal operation
as it's allowed by the spec (when done after the health tests and in a
way such that number of samples going into the conditioner is fixed).
Also, under normal operating conditions, noise source samples arrive at
very low rate and dropping bits should not be needed.

However, verifying that the `correct` entropy bits are dropped is hard
and seems impossible for our current DV environment as it requires to
very accurately model the hardware pipeline which is undesirable. Thus,
the safest approach is to add this new distribution FIFO and tune its
depth parameter to handle potential backpressure from the conditioner
such that dropping bits is not necessary.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>