[adc_ctrl] Fix low-power sampling mode
The ADC_CTRL FSM has a low power sensing mode where ADC samples are
taken at a low sampling rate. If a match is seen, it transitions to
normal sampling mode in order to confirm the match.

Unfortunately, if it does not confirm the match and the chip is in a
low-power (sleep) state, the ADC_CTRL FSM stays in normal sampling mode
and does not wake up the chip. This is not ideal from a power
perspective, since in normal sampling mode, the average power
consumption is much higher due to regular ADC sampling.

This patch fixes that and introduces a condition in the normal sampling
mode that automatically sends the FSM back to low power sampling if a
match cannot be confirmed in the normal sampling mode (in case the
ADC_CTRL is configured do use low power sampling).

Fixes #13725

Signed-off-by: Michael Schaffner <msf@opentitan.org>