[usbdev,dv] Remove nonexistent tests from testplan
A testpoint in the testplan should (eventually) be satisfied by one or
more tests that actually perform the check in question. When the
testplan was generated, the author guessed a name that should be used
for a single test for each testpoint.

This is a bit unfortunate because the dvsim tooling complains lots,
hiding any interesting information in the lines of warnings. This
commit removes the not-yet-implemented test names from the plan.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>